Put `Affine2` and `Affine3A` on the `transform-types` feature.
