+++ Cargo.toml
@@ -64,10 +64,12 @@ harness = false
 [[bench]]
 name = "affine2"
 harness = false
+required-features = ["transform-types"]
 
 [[bench]]
 name = "affine3"
 harness = false
+required-features = ["transform-types"]
 
 [[bench]]
 name = "mat4"
