+++ tests/affine3.rs
@@ -1,6 +1,8 @@
+#[cfg(feature = "transform-types")]
 #[macro_use]
 mod support;
 
+#[cfg(feature = "transform-types")]
 macro_rules! impl_affine3_tests {
     ($t:ident, $affine3:ident, $quat:ident, $vec3:ident) => {
         use core::$t::NAN;
@@ -215,6 +217,7 @@ macro_rules! impl_affine3_tests {
     };
 }
 
+#[cfg(feature = "transform-types")]
 mod affine3a {
     use super::support::deg;
     use glam::{Affine3A, Quat, Vec3, Vec3A};
@@ -247,6 +250,7 @@ mod affine3a {
     impl_affine3_tests!(f32, Affine3A, Quat, Vec3);
 }
 
+#[cfg(feature = "transform-types")]
 mod daffine3 {
     use super::support::deg;
     use glam::{DAffine3, DQuat, DVec3};
